Refactored a little bit so it's not so messy
Wasn't sure how to do it, I need a cache for the moving average and to
rewind it and take samples from the start of the stream but...

1. a class for the entire stream processor got too large too quick, was all
cruft and no code.

2. Globals are as taboo as gotos, so I avoided those.

3. A separate cache class ended up holding half of the state and made
all the token names all long and nasty.

So I've gone with a slightly messy Reader class which creates a sequence
generator on command, but keeps the average in itself.

Still doesn't feel right but my internet connection sucks